remaining case multimedia tests working
