+++ corehq/form_processor/abstract_models.py
@@ -126,9 +126,6 @@ class AbstractCommCareCase(object):
     def is_deleted(self):
         raise NotImplementedError()
 
-    def to_xml(self, version, include_case_on_closed=False):
-        raise NotImplementedError()
-
     def dynamic_case_properties(self):
         raise NotImplementedError()
 
@@ -137,6 +134,35 @@ class AbstractCommCareCase(object):
 
     def modified_since_sync(self, sync_log):
         raise NotImplementedError
+
+    def to_xml(self, version, include_case_on_closed=False):
+        from xml.etree import ElementTree
+        from casexml.apps.phone.xml import get_case_element
+        if self.closed:
+            if include_case_on_closed:
+                elem = get_case_element(self, ('create', 'update', 'close'), version)
+            else:
+                elem = get_case_element(self, ('close'), version)
+        else:
+            elem = get_case_element(self, ('create', 'update'), version)
+        return ElementTree.tostring(elem)
+
+    def get_attachment_server_url(self, attachment_key):
+        """
+        A server specific URL for remote clients to access case attachment resources async.
+        """
+        if attachment_key in self.case_attachments:
+            from dimagi.utils import web
+            from django.core.urlresolvers import reverse
+            return "%s%s" % (web.get_url_base(),
+                             reverse("api_case_attachment", kwargs={
+                                 "domain": self.domain,
+                                 "case_id": self.case_id,
+                                 "attachment_id": attachment_key,
+                             })
+            )
+        else:
+            return None
 
 
 class AbstractLedgerValue(six.with_metaclass(ABCMeta)):
